@@ -27,24 +27,24 @@
   },
   "dependencies": {
     "@1password/docusaurus-plugin-stored-data": "0.1.0",
-    "@docusaurus/core": "3.0.1",
-    "@docusaurus/plugin-pwa": "3.0.1",
-    "@docusaurus/plugin-sitemap": "3.0.1",
-    "@docusaurus/preset-classic": "3.0.1",
-    "@docusaurus/theme-mermaid": "3.0.1",
+    "@docusaurus/core": "3.1.0",
+    "@docusaurus/plugin-pwa": "3.1.0",
+    "@docusaurus/plugin-sitemap": "3.1.0",
+    "@docusaurus/preset-classic": "3.1.0",
+    "@docusaurus/theme-mermaid": "3.1.0",
     "@fec/remark-a11y-emoji": "4.0.2",
     "@mdx-js/react": "3.0.0",
     "@octokit/core": "5.0.2",
-    "@rive-app/react-canvas": "4.5.6",
+    "@rive-app/react-canvas": "4.7.0",
     "a11y-react-emoji": "1.2.0",
-    "clsx": "2.0.0",
-    "docusaurus-plugin-openapi-docs": "3.0.0-beta.3",
-    "docusaurus-theme-openapi-docs": "3.0.0-beta.3",
+    "clsx": "2.1.0",
+    "docusaurus-plugin-openapi-docs": "3.0.0-beta.4",
+    "docusaurus-theme-openapi-docs": "3.0.0-beta.4",
     "formik": "2.4.5",
     "github-slugger": "2.0.0",
     "lodash": "4.17.21",
     "netlify-plugin-cache": "1.0.3",
-    "prism-react-renderer": "2.3.0",
+    "prism-react-renderer": "2.3.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-modal": "3.16.1",
@@ -55,33 +55,33 @@
     "rehype-katex": "7.0.0",
     "remark-kroki-plugin": "0.1.1",
     "remark-math": "6.0.0",
-    "yup": "1.3.2"
+    "yup": "1.3.3"
   },
   "devDependencies": {
-    "@commitlint/cli": "18.4.3",
-    "@commitlint/config-conventional": "18.4.3",
-    "@docusaurus/module-type-aliases": "3.0.1",
-    "@docusaurus/tsconfig": "3.0.1",
-    "@docusaurus/types": "3.0.1",
-    "@typescript-eslint/eslint-plugin": "6.14.0",
-    "@typescript-eslint/parser": "6.14.0",
+    "@commitlint/cli": "18.4.4",
+    "@commitlint/config-conventional": "18.4.4",
+    "@docusaurus/module-type-aliases": "3.1.0",
+    "@docusaurus/tsconfig": "3.1.0",
+    "@docusaurus/types": "3.1.0",
+    "@typescript-eslint/eslint-plugin": "6.17.0",
+    "@typescript-eslint/parser": "6.17.0",
     "asciinema-player": "3.6.3",
     "commitizen": "4.3.0",
-    "eslint": "8.55.0",
+    "eslint": "8.56.0",
     "eslint-config-prettier": "9.1.0",
     "eslint-plugin-formatjs": "4.11.3",
     "eslint-plugin-jsx-a11y": "6.8.0",
-    "eslint-plugin-prettier": "5.0.1",
+    "eslint-plugin-prettier": "5.1.2",
     "eslint-plugin-react": "7.33.2",
     "eslint-plugin-react-hooks": "4.6.0",
     "eslint-plugin-regexp": "2.1.2",
     "husky": "8.0.3",
     "lint-staged": "15.2.0",
-    "postcss": "8.4.32",
+    "postcss": "8.4.33",
     "prettier": "3.1.1",
     "prettier-plugin-css-order": "2.0.1",
     "react-lite-youtube-embed": "^2.4.0",
-    "svgo": "3.1.0",
+    "svgo": "3.2.0",
     "typescript": "5.3.3"
   },
   "browserslist": {
